refactor: consolidate scrappers into main-scrapper
